backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
google-cloud-aiplatform==1.66.0
PyPDF2==3.0.1
# For future GCP integrations (commented for now)
# google-cloud-speech
# google-cloud-videointelligence
# google-cloud-language
# google-cloud-texttospeech
# google-cloud-storage
# google-cloud-firestore

